既にエラー型を用意してたのに、使ってなかった
