+++ src/features/jsonSchema/scala/paramSchema.ts
@@ -12,7 +12,6 @@ import type {
   SelectParam,
   StringParam,
   StringArrayParam,
-  ParamError,
 } from "@RmmzPluginSchema/rmmz/plugin";
 import type { JSONSchemaType } from "ajv";
 import type { DiscriminatedUnionSchemaType3 } from "src/libs/templates/discriminator/discriminator3";
@@ -229,18 +228,6 @@ const STRUCT_NAME = {
   maxLength: 48,
 } as const satisfies JSONSchemaType<string>;
 
-const SCHEMA_STRUCT_ERROR = {
-  type: "object",
-  additionalProperties: false,
-  required: ["message", "code", "source", "attr"],
-  properties: {
-    message: { type: "string" },
-    code: { type: "string" },
-    source: { type: "string" },
-    attr: { type: "string" },
-  },
-} as const satisfies JSONSchemaType<ParamError>;
-
 const SCHEMA_STRUCT_REF_PARAM = {
   additionalProperties: false,
   type: "object",
@@ -252,11 +239,6 @@ const SCHEMA_STRUCT_REF_PARAM = {
     text: BASIC_TEXT,
     parent: BASIC_TEXT,
     struct: STRUCT_NAME,
-    errors: {
-      nullable: true,
-      type: "array",
-      items: SCHEMA_STRUCT_ERROR,
-    },
   },
 } as const satisfies JSONSchemaType<StructRefParam>;
 
@@ -276,11 +258,6 @@ const SCHEMA_STRUCT_ARRAY_REF_PARAM = {
     desc: BASIC_TEXT,
     text: BASIC_TEXT,
     parent: BASIC_TEXT,
-    errors: {
-      nullable: true,
-      type: "array",
-      items: SCHEMA_STRUCT_ERROR,
-    },
   },
 } as const satisfies JSONSchemaType<StructArrayRefParam>;
