+++ src/rmmz/plugin/core/params/structError.ts
@@ -1,13 +1,7 @@
-import type {
-  PrimitiveParam,
-  StructArrayRefParam,
-  StructRefParam,
-} from "./types";
+import type { PluginParam } from "./types";
 
-export const isErrorStructParam = (
-  param: PrimitiveParam,
-): param is StructRefParam | StructArrayRefParam => {
-  if (param.kind === "struct" || param.kind === "struct[]") {
+export const isErrorStructParam = (param: PluginParam) => {
+  if (param.attr.kind === "struct" || param.attr.kind === "struct[]") {
     return Array.isArray(param.errors) ? param.errors.length > 0 : false;
   }
   return false;
