+++ src/rmmz/plugin/core/params/types/arrayParamItems.ts
@@ -4,27 +4,27 @@ import type {
   PrimitiveParam,
   PluginScalarParam,
 } from "./paramUnion";
-import type { StructArrayRefParam, StructRefParam } from "./primitive";
+import type { StructRefParam, StructArrayRefParam } from "./struct";
 
 export interface PluginParam {
   name: string;
   attr: PrimitiveParam;
-  erros?: ParamError[];
+  errors?: ParamError[];
 }
 
 export interface PluginParamEx<
   T extends PrimitiveParam,
-  N extends string = string
+  N extends string = string,
 > extends PluginParam {
   name: Extract<N, string>;
   attr: T;
-  erros?: ParamError[];
+  errors?: ParamError[];
 }
 
 export type PluginParamEx2<
   S extends PluginScalarParam,
   A extends PluginArrayParamType,
-  N extends string = string
+  N extends string = string,
 > = PluginParamEx<S | A | StructRefParam | StructArrayRefParam, N>;
 
 export type StructPluginParam = PluginParamEx<
