+++ src/rmmz/plugin/core/params/types/pluginSchemaType.ts
@@ -1,10 +1,6 @@
 import type { PrimitiveParam } from "./paramUnion";
-import type {
-  BooleanParam,
-  AnyStringParam,
-  StructArrayRefParam,
-  StructRefParam,
-} from "./primitive";
+import type { BooleanParam, AnyStringParam } from "./primitive";
+import type { StructArrayRefParam, StructRefParam } from "./struct";
 
 export type PluginStructParamTypeEx<T> = {
   [K in keyof T as string]: {
@@ -34,21 +30,21 @@ export type PluginParamForbiddenTypes =
 export type PluginSchemaType<T> = T extends boolean
   ? BooleanParam
   : T extends number
-  ? Extract<PrimitiveParam, { default: number }>
-  : T extends string
-  ? Extract<Exclude<PrimitiveParam, AnyStringParam>, { default: string }>
-  : T extends number[]
-  ? Extract<PrimitiveParam, { default: number[] }>
-  : T extends string[]
-  ? Extract<PrimitiveParam, { default: string[] }>
-  : T extends null
-  ? Error_NullForbidden
-  : T extends undefined
-  ? Error_UndefinedForbidden
-  : T extends boolean[]
-  ? Error_BooleanArrayForbidden
-  : T extends object[]
-  ? StructArrayRefParam
-  : T extends object
-  ? StructRefParam
-  : never;
+    ? Extract<PrimitiveParam, { default: number }>
+    : T extends string
+      ? Extract<Exclude<PrimitiveParam, AnyStringParam>, { default: string }>
+      : T extends number[]
+        ? Extract<PrimitiveParam, { default: number[] }>
+        : T extends string[]
+          ? Extract<PrimitiveParam, { default: string[] }>
+          : T extends null
+            ? Error_NullForbidden
+            : T extends undefined
+              ? Error_UndefinedForbidden
+              : T extends boolean[]
+                ? Error_BooleanArrayForbidden
+                : T extends object[]
+                  ? StructArrayRefParam
+                  : T extends object
+                    ? StructRefParam
+                    : never;
